requirements_cloud.txt
streamlit
ultralytics
torch
numpy
plotly
Pillow